+++ pymc3/step_methods/__init__.py
@@ -13,7 +13,6 @@ from .metropolis import LaplaceProposal
 from .metropolis import PoissonProposal
 from .metropolis import MultivariateNormalProposal
 
-from .sgmcmc import SGFS, CSG
 from .gibbs import ElemwiseCategorical
 
 from .slicer import Slice